Shared Function to Confirm On Back Button Click and use in group comonent

+++ Printpress.UI/Printpress_Angular/src/app/features/order/components/order-group-add-update/order-group-add-update.component.ts
@@ -213,8 +213,11 @@ export class OrderGroupAddUpdateComponent implements OnInit {
     return true;
   }
 
-  protected onBack_Click() {
-    // Alert
+  protected async onBack_Click() {
+
+    if (! await this.dialogService.confirmOnBackButton()) {
+      return;
+    }
 
     if (!this.isEdit) {
       this.orderSharedService.deleteNewlyAddedGroup(this.groupId);
